+++ src/components/branding/index.js
@@ -1,5 +1,5 @@
 import { h } from 'preact';
-import style from './style.module.less';
+import style from './style.module.css';
 import config from '../../config.json';
 
 export default function Branding() {
