+++ src/components/code-editor/index.js
@@ -4,7 +4,7 @@ import cx from '../../lib/cx';
 import 'codemirror/mode/jsx/jsx';
 import 'codemirror/addon/comment/comment';
 import 'codemirror/lib/codemirror.css';
-import style from './style.module.less';
+import style from './style.module.css';
 
 export default class CodeEditor extends Component {
 	scratch = document.createElement('div');
