+++ src/components/controllers/page/sidebar-nav.js
@@ -1,6 +1,6 @@
 import { h, Fragment } from 'preact';
 import cx from '../../../lib/cx';
-import style from './sidebar-nav.module.less';
+import style from './sidebar-nav.module.css';
 import { getCurrentUrl } from 'preact-router';
 
 /**
