+++ src/components/controllers/tutorial/style.module.css
@@ -1,5 +1,3 @@
-@import '~style/helpers';
-
 .errorOverlayWrapper {
 	position: absolute;
 	top: 0;
@@ -188,7 +186,7 @@
 					cursor: pointer;
 					&:hover,
 					&:focus-visible {
-						box-shadow: 0 0 0 5px @color-brand-light;
+						box-shadow: 0 0 0 5px var(--color-brand-light);
 					}
 
 					&[disabled] {
@@ -213,8 +211,8 @@
 		&:not(.showCode) {
 			.tutorialWindow {
 				& > * {
-					// Fixes content region whilst keeping
-					// title & button container centered
+					/* Fixes content region whilst keeping
+					   title & button container centered */
 					&:nth-child(2) {
 						width: 100%;
 					}
@@ -332,7 +330,7 @@
 			&:hover,
 			&:focus-visible {
 				box-shadow: 0 0 1px 0.5px var(--color-btn-background),
-					0 0 0 5px @color-brand-light;
+					0 0 0 5px var(--color-brand-light);
 			}
 
 			&:before {
