+++ src/components/doc-version/index.js
@@ -1,5 +1,5 @@
 import { h } from 'preact';
-import style from './style.module.less';
+import style from './style.module.css';
 import { getCurrentUrl, route } from 'preact-router';
 import { useStore } from '../store-adapter';
