+++ src/components/edit-button/style.module.css
@@ -1,0 +1,56 @@
+.wrapper {
+	position: relative;
+	width: 100%;
+	max-width: var(--content-width);
+	margin-left: auto;
+	margin-right: auto;
+
+	&.withFallback {
+		padding-top: 4rem;
+	}
+}
+
+.edit {
+	position: absolute;
+	right: 0.8rem;
+	top: 0.8rem;
+	color: var(--color-link);
+	text-decoration-skip-ink: auto;
+	text-decoration: underline;
+	text-decoration-color: var(--color-table-border);
+	text-decoration-thickness: 1px;
+
+	&:hover,
+	&:focus {
+		text-decoration-color: currentColor;
+	}
+
+	font-size: 0.9rem;
+
+	@media (min-width: 74rem) {
+		right: 0;
+	}
+}
+
+.fallback {
+	padding-top: 4rem;
+}
+
+.fallbackInner {
+	background: var(--color-warn-bg);
+	padding: 0.75rem 1rem;
+	color: #444;
+
+	a {
+		color: var(--color-link);
+		text-decoration-skip-ink: auto;
+		text-decoration: underline;
+		text-decoration-color: var(--color-table-border);
+		text-decoration-thickness: 1px;
+
+		&:hover,
+		&:focus {
+			text-decoration-color: currentColor;
+		}
+	}
+}
