--- src/components/edit-button/style.module.less
@@ -1,46 +1,0 @@
-@import '~style/helpers';
-
-.wrapper {
-	position: relative;
-	width: 100%;
-	max-width: @content-width;
-	margin-left: auto;
-	margin-right: auto;
-
-	&.withFallback {
-		padding-top: 4rem;
-	}
-}
-
-.edit {
-	position: absolute;
-	right: 0.8rem;
-	top: 0.8rem;
-	color: var(--color-link);
-	.link;
-	font-size: 0.9rem;
-
-	@media @sidebar-break {
-		right: 1.25rem;
-		top: 1rem;
-	}
-
-	@media (min-width: 74rem) {
-		right: 0;
-	}
-}
-
-.fallback {
-	padding-top: 4rem;
-}
-
-.fallbackInner {
-	background: var(--color-warn-bg);
-	padding: 0.75rem 1rem;
-	color: #444;
-
-	a {
-		.link;
-		color: var(--color-link);
-	}
-}
