+++ src/components/header/corner.js
@@ -1,5 +1,5 @@
 import { h } from 'preact';
-import style from './corner.module.less';
+import style from './corner.module.css';
 
 export default function Corner() {
 	return (
