+++ src/components/header/style.module.css
@@ -1,5 +1,3 @@
-@import '~style/helpers';
-
 .banner {
 	background: #38235c;
 	color: white;
@@ -18,13 +16,13 @@
 .outer {
 	display: flex;
 	justify-content: center;
-	height: @header-height;
+	height: var(--header-height);
 	transition: all 250ms ease;
 	transform: translateZ(0);
 
 	.open & {
 		justify-content: space-between;
-		overflow-scrolling: touch;
+		-webkit-overflow-scrolling: touch;
 		height: var(--vh);
 	}
 }
@@ -43,7 +41,7 @@
 	flex-direction: column;
 	justify-content: center;
 
-	@media (max-width: @header-mobile-breakpoint) {
+	@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 		overflow: hidden;
 
 		&.open {
@@ -51,7 +49,7 @@
 
 			overflow: auto;
 			justify-content: space-between;
-			overflow-scrolling: touch;
+			-webkit-overflow-scrolling: touch;
 
 			nav {
 				margin-right: 0;
@@ -96,7 +94,7 @@
 		user-select: none;
 		-webkit-user-drag: none;
 
-		@media (max-width: @header-mobile-breakpoint) {
+		@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 			display: block;
 			padding: 0 23px;
 		}
@@ -106,8 +104,8 @@
 		span {
 			display: inline-block;
 			position: relative;
-			height: @header-height;
-			line-height: @header-height;
+			height: var(--header-height);
+			line-height: var(--header-height);
 			padding: 0 0.6rem;
 			min-width: 50px;
 			border: none;
@@ -120,14 +118,14 @@
 			vertical-align: top;
 			appearance: none;
 
-			@media (max-width: @header-mobile-breakpoint) {
+			@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 				min-width: 100px;
 				text-align: left;
 				padding-left: 26px;
 				display: none;
 			}
 
-			@media @sidebar-break {
+			@media (min-width: /* --sidebar-break */ 900px) {
 				padding: 0 1.1rem;
 			}
 
@@ -135,7 +133,7 @@
 				text-transform: uppercase;
 				padding: 0 15px;
 
-				@media (max-width: @header-mobile-breakpoint) {
+				@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 					display: block;
 					padding: 0;
 				}
@@ -163,7 +161,7 @@
 				background-color: rgba(255, 255, 255, 0.3);
 			}
 
-			@media (max-width: @header-mobile-breakpoint) {
+			@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 				&,
 				&.current,
 				&:hover,
@@ -193,7 +191,7 @@
 			}
 		}
 
-		@media (max-width: @header-mobile-breakpoint) {
+		@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 			& {
 				display: none;
 			}
@@ -211,7 +209,7 @@
 			top: 100%;
 			margin: 0;
 			padding: 3px 0;
-			transform: translateX(-50%) perspective(1000px) translateX(0.5px); // .5 fixes blur issue
+			transform: translateX(-50%) perspective(1000px) translateX(0.5px); /* .5 fixes blur issue */
 			background: #fff;
 			border-radius: 0 0 3px 3px;
 			box-shadow: 0 2px 5px rgba(0, 0, 0, 0.4);
@@ -272,7 +270,7 @@
 				display: inline-block;
 			}
 
-			@media (max-width: @header-mobile-breakpoint) {
+			@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 				& > a,
 				& > button {
 					color: #fff;
@@ -318,19 +316,19 @@
 }
 
 .inner {
-	max-width: @full-width;
+	max-width: 80rem;
 	display: flex;
 	flex-direction: row;
 	width: 100%;
 	justify-content: space-between;
 	align-items: flex-start;
 
-	@media (min-width: @content-width) {
+	@media (min-width: /* --content-width */ 50rem) {
 		justify-content: center;
 		justify-content: flex-start;
 	}
 
-	@media @sidebar-break {
+	@media (min-width: /* --sidebar-break */ 900px) {
 		justify-content: center;
 	}
 }
@@ -339,13 +337,13 @@
 .translation {
 	position: absolute;
 	height: 2rem;
-	top: calc(var(--vh) - 3rem); // TODO: Why does bottom not work?
+	top: calc(var(--vh) - 3rem); /* TODO: Why does bottom not work? */
 	display: flex;
 	width: 100%;
 	justify-content: space-between;
 	align-items: center;
 
-	@media (min-width: @content-width) {
+	@media (min-width: /* --content-width */ 50rem) {
 		top: initial;
 		height: 100%;
 		left: initial;
@@ -395,10 +393,10 @@
 
 .release {
 	font-size: 0.9rem;
-	padding: 0 0.5rem;
+	padding: 0 0.5rem !important;
 }
 
-// hamburger menu
+/* hamburger menu */
 .hamburger {
 	display: none;
 	position: absolute;
@@ -409,7 +407,7 @@
 	width: 20px;
 	height: 20px;
 	padding: 18px 23px 12px;
-	line-height: @header-height;
+	line-height: var(--header-height);
 	text-align: center;
 	z-index: 100;
 	opacity: 0.7;
@@ -420,7 +418,7 @@
 		opacity: 1;
 	}
 
-	@media (max-width: @header-mobile-breakpoint) {
+	@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 		display: block;
 		position: relative;
 	}
@@ -465,7 +463,7 @@
 	}
 
 	&[open] {
-		@media (max-width: @header-mobile-breakpoint) {
+		@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 			position: absolute;
 		}
 
@@ -490,7 +488,7 @@
 		left: 50% !important;
 		transform: translateX(-50%) !important;
 
-		@media (max-width: @header-mobile-breakpoint) {
+		@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 			width: 90% !important;
 		}
 	}
@@ -504,7 +502,7 @@
 	background: var(--color-brand);
 	padding-right: 0.5rem;
 
-	@media (max-width: @header-mobile-breakpoint) {
+	@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 		flex-grow: 1;
 
 		.open & {
@@ -522,7 +520,7 @@
 			left: 4rem;
 			right: 4rem;
 			top: 0;
-			height: @header-height;
+			height: var(--header-height);
 			z-index: 2;
 
 			& input {
@@ -530,7 +528,7 @@
 			}
 		}
 	}
-	@media (max-width: @header-mobile-breakpoint) {
+	@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 		&:focus-within :global(.algolia-autocomplete) {
 			left: 1rem !important;
 		}
@@ -543,22 +541,22 @@
 		background-image: url('data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 490 490" width="16" height="16"><path fill="none" stroke="%238c68cf" stroke-width="36" stroke-linecap="round" d="m280,278a153,153 0 1,0-2,2l170,170m-91-117 110,110-26,26-110-110"/></svg>');
 		background-position: 10px center;
 		background-repeat: no-repeat;
-		background-color: darken(@color-brand, 7%) !important;
-		border: 1px solid lighten(@color-brand, 5%);
-		border-bottom-color: lighten(@color-brand, 10%);
+		background-color: #58319d !important; /* --color-brand 7% darker */
+		border: 1px solid #7447c5; /* --color-brand 5% lighter */
+		border-bottom-color: #825acb; /* --color-brand 10% lighter */
 		border-radius: 5px;
 		box-shadow: inset 0 2px 5px rgba(0, 0, 0, 0.2);
 		color: var(--color-brand);
 		font-size: 100%;
-		// will-change: width;
-		// transition: width 250ms ease;
+		/* will-change: width;
+		   transition: width 250ms ease; */
 
-		@media (max-width: @header-mobile-breakpoint) {
+		@media (max-width: /* --header-mobile-breakpoint */ 50rem) {
 			width: 100%;
 			margin: 10px 0;
 		}
 
-		@media (min-width: @header-mobile-breakpoint) {
+		@media (min-width: /* --header-mobile-breakpoint */ 50rem) {
 			width: 8rem;
 		}
 
@@ -567,12 +565,12 @@
 		}
 
 		&::-webkit-input-placeholder {
-			color: lighten(@color-brand, 5%);
+			color: #7447c5; /* --color-brand 5% lighter */
 		}
 
 		&:focus,
 		&:valid {
-			@media (min-width: @header-mobile-breakpoint) and (max-width: 800px) {
+			@media (min-width: /* --header-mobile-breakpoint */ 50rem) and (max-width: 800px) {
 				width: 12rem;
 			}
