+++ src/components/sponsors/style.module.css
@@ -1,53 +1,54 @@
-@import '~style/helpers';
-
-[name='/about/we-are-using'] h1,
-[name='/about/we-are-using'] p {
+:global(.sponsors) {
 	text-align: center;
+	margin-bottom: -2rem;
+	max-width: 60rem !important;
+	padding-left: 0 !important;
+	padding-right: 0 !important;
+
+	p {
+		margin-bottom: 0;
+	}
 }
 
-html .root {
+.sponsorList {
 	display: flex;
 	flex-wrap: wrap;
 	justify-content: center;
-	margin: 4rem auto 0;
-	padding: 0;
-	max-width: 120rem;
-
-	@media @sidebar-break {
-		margin: 4rem auto 2rem;
-	}
+	align-items: center;
+	padding: 0 !important;
+	margin: 0.5rem 0 0;
 }
 
-.item {
-	display: inline-block;
+.sponsorItem {
+	flex: 1 1 auto;
 	list-style: none;
-	width: 10rem;
+	transition: all 0.3s;
+	text-align: center;
+	padding: 0.5rem 1.5rem 0;
+	margin: 0;
 
-	@media @sidebar-break {
-		width: 12.5rem;
-	}
-
-	a {
-		display: block;
-		padding: 2rem 1.25rem;
-		margin-bottom: 1rem;
-		transition: all 250ms ease;
-		text-align: center;
+	& a {
+		text-decoration: none;
+		display: inline-block;
+		color: transparent !important;
+		/* Fix IE 11 border */
 		background: none !important;
-		padding: 1rem;
 
-		&:hover img {
+		&:hover img,
+		&:hover svg {
 			filter: grayscale(0);
 			opacity: 1;
 		}
 
-		:global(.dark) &:hover img {
+		:global(.dark) &:hover img,
+		:global(.dark) &:hover svg {
 			filter: invert(0) hue-rotate(0deg) grayscale(0)
 				drop-shadow(0 0 0.5px #fff) drop-shadow(0 5px 10px #000)
 				drop-shadow(0 0 20px rgba(255, 255, 255, 0.5));
 		}
 		@media (prefers-color-scheme: dark) {
-			&:hover img {
+			&:hover img,
+			&:hover svg {
 				filter: invert(0) hue-rotate(0deg) grayscale(0)
 					drop-shadow(0 0 0.5px #fff) drop-shadow(0 5px 10px #000)
 					drop-shadow(0 0 20px rgba(255, 255, 255, 0.5));
@@ -55,15 +56,17 @@ html .root {
 		}
 	}
 
-	img {
-		height: 4.5rem;
+	img,
+	svg {
+		height: 3rem;
+		width: auto;
 		display: block;
 		margin: 0 auto 1rem;
 		opacity: 0.6;
 		transition: all 250ms ease;
 		filter: grayscale(1);
 
-		// Invert based on luminosity
+		/* Invert based on luminosity */
 		@media (prefers-color-scheme: dark) {
 			filter: invert(0.862745) hue-rotate(180deg) grayscale(1);
 		}
@@ -72,9 +75,4 @@ html .root {
 			opacity: 1;
 		}
 	}
-}
-
-:global(.breaker) {
-	margin: 0 !important;
-	max-width: none !important;
 }
