+++ src/components/table-of-contents/index.js
@@ -1,6 +1,6 @@
 import { h } from 'preact';
 import cx from '../../lib/cx';
-import style from './style.module.less';
+import style from './style.module.css';
 import { useStore } from '../store-adapter';
 import { useMemo, useRef } from 'preact/hooks';
