+++ src/components/we-are-using/index.js
@@ -1,4 +1,4 @@
-import styles from './style.module.less';
+import styles from './style.module.css';
 
 // Add your company to this array. The logo should be placed
 // in `src/assets/we-are-using/my-logo.svg`.
