+++ src/components/we-are-using/style.module.css
@@ -1,54 +1,47 @@
-:global(.sponsors) {
+[name='/about/we-are-using'] h1,
+[name='/about/we-are-using'] p {
 	text-align: center;
-	margin-bottom: -2rem;
-	max-width: 60rem !important;
-	padding-left: 0 !important;
-	padding-right: 0 !important;
-
-	p {
-		margin-bottom: 0;
-	}
 }
 
-.sponsorList {
+html .root {
 	display: flex;
 	flex-wrap: wrap;
 	justify-content: center;
-	align-items: center;
-	padding: 0 !important;
-	margin: 0.5rem 0 0;
+	margin: 4rem auto 0;
+	padding: 0;
+	max-width: 120rem;
 }
 
-.sponsorItem {
-	flex: 1 1 auto;
+.item {
+	display: inline-block;
 	list-style: none;
-	transition: all 0.3s;
-	text-align: center;
-	padding: 0.5rem 1.5rem 0;
-	margin: 0;
+	width: 10rem;
+
+	@media (min-width: 900px) {
+		width: 12.5rem;
+	}
 
 	a {
-		text-decoration: none;
-		display: inline-block;
-		color: transparent !important;
-		// Fix IE 11 border
+		display: block;
+		padding: 2rem 1.25rem;
+		margin-bottom: 1rem;
+		transition: all 250ms ease;
+		text-align: center;
 		background: none !important;
+		padding: 1rem;
 
-		&:hover img,
-		&:hover svg {
+		&:hover img {
 			filter: grayscale(0);
 			opacity: 1;
 		}
 
-		:global(.dark) &:hover img,
-		:global(.dark) &:hover svg {
+		:global(.dark) &:hover img {
 			filter: invert(0) hue-rotate(0deg) grayscale(0)
 				drop-shadow(0 0 0.5px #fff) drop-shadow(0 5px 10px #000)
 				drop-shadow(0 0 20px rgba(255, 255, 255, 0.5));
 		}
 		@media (prefers-color-scheme: dark) {
-			&:hover img,
-			&:hover svg {
+			&:hover img {
 				filter: invert(0) hue-rotate(0deg) grayscale(0)
 					drop-shadow(0 0 0.5px #fff) drop-shadow(0 5px 10px #000)
 					drop-shadow(0 0 20px rgba(255, 255, 255, 0.5));
@@ -56,17 +49,15 @@
 		}
 	}
 
-	img,
-	svg {
-		height: 3rem;
-		width: auto;
+	img {
+		height: 4.5rem;
 		display: block;
 		margin: 0 auto 1rem;
 		opacity: 0.6;
 		transition: all 250ms ease;
 		filter: grayscale(1);
 
-		// Invert based on luminosity
+		/* Invert based on luminosity */
 		@media (prefers-color-scheme: dark) {
 			filter: invert(0.862745) hue-rotate(180deg) grayscale(1);
 		}
@@ -75,8 +66,9 @@
 			opacity: 1;
 		}
 	}
+}
 
-	// & + .sponsorItem {
-	// 	margin-left: 3rem;
-	// }
+:global(.breaker) {
+	margin: 0 !important;
+	max-width: none !important;
 }
