+++ src/style/buttons.css
@@ -3,7 +3,6 @@
 	min-width: 13em;
 	padding: 0.5rem 1.5rem;
 	margin: 0 1rem;
-	box-shadow: 0 0.125rem 0 rgba(var(--color-brand), 0.3);
 	border: 0.2rem solid var(--color-btn-background);
 	border-radius: 0.25rem;
 	text-transform: uppercase;
@@ -44,13 +43,11 @@
 	display: inline-block;
 	padding: 0.25rem 1rem;
 	font-size: 1rem;
-	box-shadow: 0 0.125rem 0 rgba(var(--color-brand), 0.3);
 	border: 0.2rem solid var(--color-brand);
 	border-radius: 0.25rem;
 	text-decoration: none;
 	white-space: nowrap;
 	background: var(--color-brand);
-	text-decoration: none;
 	color: white;
 	transform: translate3d(0, 0, 0);
 	transition: background-color 0.3s, border-color 0.3s, color 0.3s;
