--- src/style/helpers.less
@@ -1,2 +1,0 @@
-@import 'variables';
-@import 'mixins';
