+++ src/style/home.css
@@ -1,0 +1,188 @@
+main .markup {
+	.tagline {
+		display: block;
+		font-size: 1.5rem;
+		font-weight: 300;
+		line-height: 1.2;
+		color: white;
+		text-align: center;
+		margin: 3rem 0 3.5rem;
+	}
+
+	.home-top {
+		margin: 60px auto;
+		max-width: 984px;
+
+		& h1 {
+			font-size: 7.3vw;
+			text-align: center;
+			font-weight: lighter;
+			letter-spacing: 0.02em;
+
+			@media (min-width: 1024px) {
+				font-size: 400%;
+			}
+		}
+	}
+
+	.intro-buttons {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		margin-bottom: 3rem;
+
+		@media (min-width: 600px) {
+			flex-direction: row-reverse;
+			justify-content: center;
+			align-items: initial;
+			margin-bottom: 4.5rem;
+		}
+	}
+
+	/* Styles for a typical media/flag layout where you have one image on the left
+	or right and text in the other column. */
+	.home-section {
+		display: flex;
+		flex-direction: column;
+		max-width: 61.5rem !important; /* TODO: Fix markup page styles */
+
+		@media (min-width: 700px) {
+			flex-direction: row;
+		}
+
+		& + .home-section {
+			margin: 1.6rem auto;
+
+			@media (min-width: 700px) {
+				margin: 3.75rem auto;
+			}
+		}
+
+		/* The following nodes are created inside the markdown files.
+		To make the authoring experience easier we select and style them
+		based on the type instead of forcing content creators to add classes. */
+		& h2 {
+			font-size: 220%;
+			letter-spacing: 0.01em;
+			font-weight: 300;
+			text-align: center;
+			line-height: 1.2;
+			margin-bottom: 0.5em;
+
+			@media (min-width: 700px) {
+				margin-top: 0;
+				text-align: left;
+			}
+		}
+
+		/* TODO: Inline SVGs instead of using images to allow more styling options
+		via CSS. We could play with subtle gradients for the stroke-color, etc. */
+		& img {
+			display: block;
+			width: 25%;
+			align-self: center;
+			height: auto; /* override intrinsic size */
+			max-height: 13rem; /* IE11 fix */
+
+			@media (min-width: 700px) {
+				width: 100%;
+				flex: 22% 0 0;
+				margin-right: 4%;
+			}
+		}
+
+		@media (min-width: 700px) {
+			/* Variant: Right aligned image */
+			&:nth-of-type(2n + 1) {
+				flex-direction: row-reverse;
+
+				& img {
+					margin-right: 0;
+					margin-left: 4%;
+				}
+			}
+		}
+	}
+
+	.home-split {
+		margin-bottom: 2rem;
+		display: flex;
+		flex-direction: column;
+		max-width: 61.5rem !important; /* TODO: Fix markup page styles */
+
+		@media (min-width: 900px) {
+			flex-direction: row;
+		}
+
+		& h2 {
+			font-weight: inherit;
+		}
+
+		> div {
+			float: left;
+			box-sizing: border-box;
+			width: 40%;
+
+			@media (min-width: 900px) {
+				padding: 10px;
+
+				&:first-child {
+					width: 60%;
+				}
+			}
+		}
+
+		@media (max-width: 900px) {
+			max-width: 600px;
+			margin: auto;
+
+			> div:last-child > h2 {
+				margin-bottom: 0;
+				font-size: 20px;
+				border-bottom: none;
+			}
+
+			& pre {
+				margin: 10px 0;
+			}
+
+			> div {
+				float: none !important;
+				width: auto !important;
+			}
+		}
+	}
+
+	.home-demo {
+		padding: 20px;
+		background: #f8f8f8;
+		border: 2px solid #ddd;
+
+		& label {
+			color: #444;
+
+			& span:first-child {
+				display: block;
+			}
+		}
+
+		& input,
+		& button {
+			margin: 0 5px 0 0.5rem;
+			padding: 5px 10px;
+			border: 1px solid #673ab8;
+			border-radius: 3px;
+			font-size: 100%;
+		}
+
+		& button {
+			background: #673ab8;
+			color: #fff;
+		}
+
+		& input {
+			box-sizing: border-box;
+			width: 75%;
+		}
+	}
+}
