--- src/style/home.less
@@ -1,9 +1,0 @@
-.tagline {
-	display: block;
-	font-size: 1.5rem;
-	font-weight: 300;
-	line-height: 1.2;
-	color: white;
-	text-align: center;
-	margin: 3rem 0 3.5rem;
-}
