+++ src/style/index.css
@@ -1,0 +1,139 @@
+@import './buttons.css';
+@import './markdown.css';
+@import './gh-btn.css';
+@import './list-view.css';
+@import './prism.css';
+@import './variables.css';
+@import './home.css';
+
+html,
+body {
+	padding: 0;
+	margin: 0;
+	background: var(--color-page-bg);
+	color: var(--color-text);
+	font: 16px/1.65 system-ui, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
+		'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+	font-weight: 400;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+}
+
+body {
+	/* Sets overflowing background color when scrolled past the page in MacOS or in iOS. */
+	background: var(--color-brand);
+}
+
+* {
+	box-sizing: border-box;
+	-webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+}
+
+ol,
+ul {
+	margin-left: 0;
+}
+
+select {
+	font-size: 100%;
+}
+
+a {
+	text-decoration: none;
+
+	.white {
+		color: #fff;
+	}
+}
+
+h1,
+h2 {
+	color: var(--color-heading);
+}
+
+pre,
+code {
+	font-family: 'source-code-pro', Menlo, Consolas, Monaco, 'Andale Mono',
+		'Courier New', monospace;
+	font-size: 0.9rem;
+}
+
+div.highlight-container {
+	padding: 0 !important;
+	position: relative;
+	margin-bottom: 2.5rem;
+
+	pre.highlight {
+		position: relative;
+		border: none;
+		line-height: 1.5;
+		padding: 20px;
+	}
+
+	a.repl-link {
+		font-size: 13px;
+		position: absolute;
+		right: 0;
+		top: 0;
+		padding: 3px 7px;
+		margin: 0;
+		color: #fff;
+		border: none;
+		background: var(--color-repl-bg);
+		border-radius: 0 3px 0 3px;
+		text-align: center;
+		text-decoration: none;
+
+		&:hover {
+			background: var(--color-repl-hover);
+		}
+	}
+}
+
+#app {
+	min-height: 100%;
+	background: var(--color-page-bg);
+
+	& > main {
+		min-height: 95%;
+		padding-top: var(--header-and-banner-height);
+		display: block; /* Fix IE11 layout */
+	}
+}
+
+progress-bar {
+	position: fixed;
+	left: 0;
+	top: -5px;
+	width: 100%;
+	height: 3px;
+	background: fade(darken(#f2777a, 10%), 50%);
+	transform: translateY(0);
+	transition: transform 1s ease, opacity 1s ease;
+	opacity: 0;
+	z-index: 1000;
+	&::after {
+		content: '';
+		display: block;
+		height: 100%;
+		width: 75%;
+		background: #f2777a;
+		transform: translateX(-100%);
+		will-change: transform;
+	}
+	&[showing] {
+		transform: translateY(5px);
+		opacity: 1;
+		&::after {
+			animation: fly 1s linear forwards infinite;
+		}
+	}
+}
+@keyframes fly {
+	0% {
+		transform: translateX(-100%);
+	}
+	100% {
+		transform: translateX(150%);
+	}
+}
