--- src/style/index.less
@@ -1,313 +1,0 @@
-@import './helpers';
-@import './reset';
-@import './buttons';
-@import './markdown';
-@import './gh-btn';
-@import './list-view';
-@import './prism.css';
-@import './variables';
-@import './home';
-
-html,
-body {
-	padding: 0;
-	margin: 0;
-	background: var(--color-page-bg);
-	color: var(--color-text);
-	font: 16px/1.65 system-ui, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
-		'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
-	font-weight: 400;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
-}
-
-body {
-	// Sets overflowing background color when scrolled past the page in
-	// MacOS or in iOS.
-	background: var(--color-brand);
-}
-
-* {
-	box-sizing: border-box;
-	-webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-}
-
-a {
-	text-decoration: none;
-}
-
-a.white {
-	color: #fff;
-}
-
-h1,
-h2 {
-	color: var(--color-heading);
-}
-
-pre,
-code {
-	font-family: 'source-code-pro', Menlo, Consolas, Monaco, 'Andale Mono',
-		'Courier New', monospace;
-	font-size: 0.9rem;
-}
-
-div.highlight-container {
-	padding: 0 !important;
-	position: relative;
-	margin-bottom: 2.5rem;
-
-	pre.highlight {
-		position: relative;
-		border: none;
-		line-height: 1.5;
-		padding: 20px;
-	}
-
-	a.repl-link {
-		font-size: 13px;
-		position: absolute;
-		right: 0;
-		top: 0;
-		padding: 3px 7px;
-		margin: 0;
-		color: #fff;
-		border: none;
-		background: var(--color-repl-bg);
-		border-radius: 0 3px 0 3px;
-		text-align: center;
-		text-decoration: none;
-
-		&:hover {
-			background: var(--color-repl-hover);
-		}
-	}
-}
-
-#app {
-	min-height: 100%;
-	background: var(--color-page-bg);
-
-	& > main {
-		min-height: 95%;
-		padding-top: @header-and-banner-height;
-		display: block; // Fix IE11 layout
-	}
-}
-
-progress-bar {
-	position: fixed;
-	left: 0;
-	top: -5px;
-	width: 100%;
-	height: 3px;
-	background: fade(darken(#f2777a, 10%), 50%);
-	transform: translateY(0);
-	transition: transform 1s ease, opacity 1s ease;
-	opacity: 0;
-	z-index: 1000;
-	&::after {
-		content: '';
-		display: block;
-		height: 100%;
-		width: 75%;
-		background: #f2777a;
-		transform: translateX(-100%);
-		will-change: transform;
-	}
-	&[showing] {
-		transform: translateY(5px);
-		opacity: 1;
-		&::after {
-			animation: fly 1s linear forwards infinite;
-		}
-	}
-}
-@keyframes fly {
-	0% {
-		transform: translateX(-100%);
-	}
-	100% {
-		transform: translateX(150%);
-	}
-}
-
-main .markup {
-	.home-top {
-		margin: 60px auto;
-		max-width: 984px;
-
-		h1 {
-			font-size: 7.3vw;
-			text-align: center;
-			font-weight: lighter;
-			letter-spacing: 0.02em;
-
-			@media (min-width: 1024px) {
-				font-size: 400%;
-			}
-		}
-	}
-
-	.intro-buttons {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		margin-bottom: 3rem;
-
-		@media (min-width: 600px) {
-			flex-direction: row-reverse;
-			justify-content: center;
-			align-items: initial;
-			margin-bottom: 4.5rem;
-		}
-	}
-
-	.home-demo {
-		padding: 20px;
-		background: #f8f8f8;
-		border: 2px solid #ddd;
-
-		label {
-			color: #444;
-			span:first-child {
-				display: block;
-			}
-		}
-
-		input,
-		button {
-			margin: 0 5px 0 0.5rem;
-			padding: 5px 10px;
-			border: 1px solid #673ab8;
-			border-radius: 3px;
-			font-size: 100%;
-		}
-		button {
-			background: #673ab8;
-			color: #fff;
-		}
-		input {
-			box-sizing: border-box;
-			width: 75%;
-		}
-	}
-}
-
-.home-split {
-	padding: 0 10px;
-	margin-bottom: 2rem;
-	display: flex;
-	flex-direction: column;
-	max-width: @home-width !important; // TODO: Fix markup page styles
-
-	@media @sidebar-break {
-		padding: 0 10px;
-		flex-direction: row;
-	}
-
-	h2 {
-		font-weight: inherit;
-	}
-
-	> div {
-		float: left;
-		box-sizing: border-box;
-		width: 40%;
-
-		@media @sidebar-break {
-			padding: 10px;
-
-			&:first-child {
-				width: 60%;
-			}
-		}
-	}
-
-	@media (max-width: 900px) {
-		max-width: 600px;
-		margin: auto;
-
-		> div:last-child > h2 {
-			margin-bottom: 0;
-			font-size: 20px;
-			border-bottom: none;
-		}
-
-		pre {
-			margin: 10px 0;
-		}
-
-		> div {
-			float: none !important;
-			width: auto !important;
-		}
-	}
-}
-
-@break-s: ~'(min-width: 700px)';
-
-// Styles for a typical media/flag layout where you have one image on the left
-// or right and text in the other column.
-.home-section {
-	display: flex;
-	flex-direction: column;
-	max-width: @home-width !important; // TODO: Fix markup page styles
-
-	@media @break-s {
-		flex-direction: row;
-	}
-
-	& + .home-section {
-		margin: 1.6rem 0;
-
-		@media @break-s {
-			margin: 3.75rem 0;
-		}
-	}
-
-	// The following nodes are created inside the markdown files.
-	// To make the authoring experience easier we select and style them
-	// based on the type instead of forcing content creators to add classes.
-	h2 {
-		font-size: 220%;
-		letter-spacing: 0.01em;
-		font-weight: 300;
-		text-align: center;
-		line-height: 1.2;
-		margin-bottom: 0.5em;
-
-		@media @break-s {
-			margin-top: 0;
-			text-align: left;
-		}
-	}
-
-	// TODO: Inline SVGs instead of using images to allow more styling options
-	// via CSS. We could play with subtle gradients for the stroke-color, etc.
-	img {
-		display: block;
-		width: 25%;
-		align-self: center;
-		height: auto; // override intrinsic size
-		max-height: 13rem; // IE11 fix
-
-		@media @break-s {
-			width: 100%;
-			flex: 22% 0 0;
-			margin-right: 4%;
-		}
-	}
-
-	@media @break-s {
-		// Variant: Right aligned image
-		&:nth-of-type(2n + 1) {
-			flex-direction: row-reverse;
-
-			img {
-				margin-right: 0;
-				margin-left: 4%;
-			}
-		}
-	}
-}
