--- src/style/mixins.less
@@ -1,39 +1,0 @@
-.fill() {
-	position: absolute;
-	left: 0;
-	top: 0;
-	width: 100%;
-	height: 100%;
-}
-
-.link() {
-	text-decoration-skip-ink: auto;
-	text-decoration: underline;
-	text-decoration-color: var(--color-table-border);
-	text-decoration-thickness: 1px;
-	
-	&:hover, &:focus {
-		text-decoration-color: currentColor;
-	}
-}
-
-.scroll() {
-	overflow: auto;
-	overflow-scrolling: touch;
-
-	& > .inner {
-		position: relative;
-		transform: translateZ(0);
-		overflow: hidden;
-	}
-}
-
-.fadeOut() {
-	visibility: hidden;
-	opacity: 0;
-}
-
-.fadeIn() {
-	visibility: visible;
-	opacity: 1;
-}
