--- src/style/reset.less
@@ -1,8 +1,0 @@
-ol,
-ul {
-	margin-left: 0;
-}
-
-select {
-	font-size: 100%;
-}
